pyproject.toml
[project]
name = "agentia"
version = "0.0.19"
description = "Ergonomic LLM Agents"
authors = [{ name = "Wenyu Zhao", email = "wenyuzhaox@gmail.com" }]
requires-python = ">=3.12"
readme = "README.md"
license = "MIT"
keywords = [
    "chatgpt",
    "gpt",
    "llm",
    "agent",
    "tools",
    "openai",
    "openrouter",
    "deepseek",
]
dependencies = [
    "filelock>=3.18.0",
    "mcp>=1.12.0",
    "nest-asyncio>=1.6.0",
    "openai>=1.108.1,<2",
    "pillow>=11.2.1",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.0,<2",
    "websockets>=15.0.1",
]

[project.urls]
Repository = "https://github.com/wenyuzhao/agentia"

[dependency-groups]
# Development tools
dev = [
    "pyright>=1.1.361,<2",
    "pytest>=7.3.2,<8",
    "pytest-asyncio>=0.21.1,<0.22",
]
# Include all plugins
plugins = ["markdownify>=1.1.0", "tavily-python>=0.7.12"]
# Include all plugins, the web app, and the API server
all = ["agentia[plugins]", "agentia[dev]"]

[tool.hatch.build.targets.sdist]
include = ["agentia"]

[tool.hatch.build.targets.wheel]
include = ["agentia"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv.workspace]
members = ["agentia-app"]
